+++ src/com/fuwei/util/NumberUtil.java
@@ -6,7 +6,7 @@ public class NumberUtil {
 		for(int i = 0 ; i < fixed ; ++i){
 			sep *= 10;
 		}
-		return (double) (Math.round(value*100)/100.0);
+		return (Math.round(value*100)/100.0);
 	}
 //	public static double percent(double value,int fixed){
 //		double temp = value * 100;
